requirements.txt
langchain_groq
langchain
langchain_community
python-dotenv
ipykernel
transformers
flask
streamlit
bs4
langchain_openai
faiss-cpu